cip68_example_implement/off_chain/requirements.txt
pycardano>=0.11.0
python-dotenv>=1.0.0
mnemonic>=0.20
cbor2>=5.4.6
